refactor: one redb corruption classifier for the workspace (Refs #5063)
Five crates each carried a byte-identical copy of the four-arm match that
decides whether an unopenable redb file may be quarantined
(UpgradeRequired | RepairAborted | Storage(Corrupted) | Storage(Io(InvalidData))),
so a safety change to that decision had to land five times.

trusty-common gains `redb_open` behind a new light `redb-open` feature gating
only `dep:redb`. The classifier used to sit under `memory-core`, which pulls in
usearch, git2 and a bundled ORT embedder — the cost no consumer would pay for a
predicate, and the reason the copies existed. `memory-core` enables the new
feature and `memory_core::store::redb_open` re-exports every item, so no
existing path moves.

Routed: trusty-common, trusty-analyze (is_format_obsolete), trusty-agents
(is_incompatible), trusty-review (redb_error_is_incompatible_format), and
trusty-search (is_incompatible_corpus_format). trusty-search's
backup_incompatible_corpus and INCOMPATIBLE_CORPUS_SUFFIX also route through
the shared quarantine-path helper, which it duplicated verbatim including the
numbered anti-clobber rule.

Each store's recovery POLICY stays where it is, with a comment naming why it
diverges: trusty-search fails closed on genuine corruption (#4227) and opens
with a tuned page-cache size, trusty-review serialises recovery behind a
sidecar lock (#5064), trusty-analyze takes a caller-supplied suffix, and
trusty-agents renames to a fixed suffix with anyhow context. No call site
changes behaviour.

`classifier_pins_the_four_recoverable_arms` states the intended verdict for
every constructible DatabaseError variant independently of the classifier and
asserts the true-set is exactly those four by name. Verified both directions:
adding a fifth arm and dropping one each turn it red.
